Redid all my code.
Making my specific scripts better for when the project gets massive.

+++ Assets/Enemies/Scripts/Enemy.cs
@@ -3,15 +3,15 @@ using UnityEngine;
 public class Enemy : MonoBehaviour, IDamageTaken
 {
     [SerializeField]
-    protected int health;
+    protected float health;
     [SerializeField]
-    protected int damage;
+    protected float damage;
     IGetHealth getHealth;
     
 
-    public void takeDamage(int damageTaken, GameObject weapon)
+    public void takeDamage(float damageTaken, GameObject weapon)
     {
-      //  Debug.Log($"Enemy has been hit! {damageTaken} damage has been dealt");
+
         health -= damageTaken;
         getHealth = weapon.GetComponent<IGetHealth>();
         if (getHealth != null)
@@ -20,7 +20,7 @@ public class Enemy : MonoBehaviour, IDamageTaken
             Debug.Log(weapon);
         if (health <= 0)
             enemyDeath();
-      //  Debug.Log($"Enemy has {health} left");
+
     }
